split up opttoy.go
