+++ v3/build.go
@@ -1,0 +1,266 @@
+package v3
+
+import "github.com/cockroachdb/cockroach/pkg/sql/parser"
+
+func build(stmt parser.Statement) *expr {
+	switch stmt := stmt.(type) {
+	case *parser.Select:
+		return buildSelect(stmt)
+	case *parser.ParenSelect:
+		return buildSelect(stmt.Select)
+	default:
+		unimplemented("%T", stmt)
+		return nil
+	}
+}
+
+func buildTable(table parser.TableExpr) *expr {
+	switch source := table.(type) {
+	case *parser.NormalizableTableName:
+		return &expr{
+			op:   scanOp,
+			body: source,
+		}
+	case *parser.AliasedTableExpr:
+		return buildTable(source.Expr)
+	case *parser.ParenTableExpr:
+		return buildTable(source.Expr)
+	case *parser.JoinTableExpr:
+		result := &expr{
+			op: innerJoinOp,
+			children: []*expr{
+				buildTable(source.Left),
+				buildTable(source.Right),
+			},
+			inputCount: 2,
+		}
+
+		switch cond := source.Cond.(type) {
+		case *parser.OnJoinCond:
+			result.addFilter(buildExpr(cond.Expr))
+
+		case parser.NaturalJoinCond:
+			result.body = cond
+		case *parser.UsingJoinCond:
+			result.body = cond
+
+		default:
+			unimplemented("%T", source.Cond)
+		}
+		return result
+	case *parser.Subquery:
+		return build(source.Select)
+
+	default:
+		unimplemented("%T", table)
+		return nil
+	}
+}
+
+func buildExpr(pexpr parser.Expr) *expr {
+	switch t := pexpr.(type) {
+	case *parser.ParenExpr:
+		return buildExpr(t.Expr)
+
+	case *parser.AndExpr:
+		return &expr{
+			op: andOp,
+			children: []*expr{
+				buildExpr(t.Left),
+				buildExpr(t.Right),
+			},
+			inputCount: 2,
+		}
+	case *parser.OrExpr:
+		return &expr{
+			op: orOp,
+			children: []*expr{
+				buildExpr(t.Left),
+				buildExpr(t.Right),
+			},
+			inputCount: 2,
+		}
+	case *parser.NotExpr:
+		return &expr{
+			op: notOp,
+			children: []*expr{
+				buildExpr(t.Expr),
+			},
+			inputCount: 1,
+		}
+
+	case *parser.BinaryExpr:
+		return &expr{
+			op: binaryOpMap[t.Operator],
+			children: []*expr{
+				buildExpr(t.Left),
+				buildExpr(t.Right),
+			},
+			inputCount: 2,
+		}
+	case *parser.ComparisonExpr:
+		return &expr{
+			op: comparisonOpMap[t.Operator],
+			children: []*expr{
+				buildExpr(t.Left),
+				buildExpr(t.Right),
+			},
+			inputCount: 2,
+		}
+	case *parser.UnaryExpr:
+		return &expr{
+			op: unaryOpMap[t.Operator],
+			children: []*expr{
+				buildExpr(t.Expr),
+			},
+			inputCount: 1,
+		}
+
+	case parser.UnqualifiedStar:
+		return &expr{
+			op:   variableOp,
+			body: t,
+		}
+	case parser.UnresolvedName:
+		return &expr{
+			op:   variableOp,
+			body: t,
+		}
+	case *parser.NumVal:
+		return &expr{
+			op:   constOp,
+			body: t,
+		}
+
+	case *parser.ExistsExpr:
+		return &expr{
+			op: existsOp,
+			children: []*expr{
+				buildExpr(t.Subquery),
+			},
+			inputCount: 1,
+		}
+
+	case *parser.Subquery:
+		return build(t.Select)
+
+	default:
+		unimplemented("%T", pexpr)
+		return nil
+	}
+}
+
+func buildSelect(stmt *parser.Select) *expr {
+	switch t := stmt.Select.(type) {
+	case *parser.SelectClause:
+		// TODO: stmt.Limit
+		return buildSelectClause(t, stmt.OrderBy)
+	case *parser.UnionClause:
+		return buildUnion(t)
+	// TODO: handle other stmt.Select types.
+	default:
+		unimplemented("%T", stmt.Select)
+	}
+	return nil
+}
+
+func buildSelectClause(clause *parser.SelectClause, orderBy parser.OrderBy) *expr {
+	var result *expr
+	if clause.From != nil {
+		var inputs []*expr
+		for _, table := range clause.From.Tables {
+			inputs = append(inputs, buildTable(table))
+		}
+		if len(inputs) == 1 {
+			result = inputs[0]
+		} else {
+			result = &expr{
+				op:         innerJoinOp,
+				children:   inputs,
+				inputCount: int16(len(inputs)),
+				body:       parser.NaturalJoinCond{},
+			}
+		}
+		if clause.Where != nil {
+			result = &expr{
+				op: selectOp,
+				children: []*expr{
+					result,
+				},
+				inputCount: 1,
+			}
+			result.addFilter(buildExpr(clause.Where.Expr))
+		}
+	}
+
+	if clause.GroupBy != nil {
+		result = &expr{
+			op:         groupByOp,
+			children:   []*expr{result},
+			inputCount: 1,
+		}
+		if clause.Having != nil {
+			result = &expr{
+				op: selectOp,
+				children: []*expr{
+					result,
+				},
+				inputCount: 1,
+			}
+			result.addFilter(buildExpr(clause.Having.Expr))
+		}
+	}
+
+	if len(clause.Exprs) > 0 {
+		result = &expr{
+			op: projectOp,
+			children: []*expr{
+				result,
+			},
+			inputCount: 1,
+		}
+		for _, expr := range clause.Exprs {
+			result.addProjection(buildExpr(expr.Expr))
+		}
+	}
+
+	if clause.Distinct {
+		result = &expr{
+			op:         distinctOp,
+			children:   []*expr{result},
+			inputCount: 1,
+		}
+	}
+
+	// TODO: order by is not a relational expression, but instead a required
+	// property on the output.
+	if orderBy != nil {
+		result = &expr{
+			op:         orderByOp,
+			children:   []*expr{result},
+			inputCount: 1,
+			body:       orderBy,
+		}
+	}
+
+	return result
+}
+
+func buildUnion(clause *parser.UnionClause) *expr {
+	op := unionOp
+	switch clause.Type {
+	case parser.UnionOp:
+	case parser.IntersectOp:
+		op = intersectOp
+	case parser.ExceptOp:
+		op = exceptOp
+	}
+	return &expr{
+		op: op,
+		children: []*expr{
+			buildSelect(clause.Left),
+			buildSelect(clause.Right),
+		},
+		inputCount: 2,
+	}
+}
