+++ v3/distinct.go
@@ -1,0 +1,7 @@
+package v3
+
+func init() {
+	operatorTab[distinctOp] = operatorInfo{
+		name: "distinct",
+	}
+}
