+++ v3/group.go
@@ -1,0 +1,7 @@
+package v3
+
+func init() {
+	operatorTab[groupByOp] = operatorInfo{
+		name: "groupBy",
+	}
+}
